Improve player list loading
Instead of reloading each time, now we have a LiveData which would notify observers of any changes (add, remove, delete) to the players stored in the repo. This is the observed and the data updates are mapped to the correct UI model.

Also added an either type to encapsulate the state of the update from the various repos.